add delete option for services
